backend/requirements.txt
Flask
Flask-PyMongo
Flask-Cors
python-dotenv
opencv-python
numpy